+++ src/main/CmdCardReadBinary.h
@@ -57,7 +57,7 @@ public:
      */
     CmdCardReadBinary(const CalypsoCardClass calypsoCardClass,
                       const uint8_t sfi,
-                      const int offset,
+                      const uint8_t offset,
                       const uint8_t length);
 
     /**
@@ -82,7 +82,7 @@ public:
      * @return The offset.
      * @since 2.1.0
      */
-    int getOffset() const;
+    uint8_t getOffset() const;
 
     /**
      * {@inheritDoc}
@@ -111,7 +111,7 @@ private:
     /**
      *
      */
-    const int mOffset;
+    const uint8_t mOffset;
 
     /**
      *
