+++ src/main/CmdCardUpdateRecord.cpp
@@ -36,7 +36,7 @@ const std::map<const int, const std::shared_ptr<StatusProperties>>
 
 CmdCardUpdateRecord::CmdCardUpdateRecord(const CalypsoCardClass calypsoCardClass,
                                          const uint8_t sfi,
-                                         const int recordNumber,
+                                         const uint8_t recordNumber,
                                          const std::vector<uint8_t>& newRecordData)
 : AbstractCardCommand(mCommand),
   mSfi(sfi),
@@ -67,12 +67,12 @@ bool CmdCardUpdateRecord::isSessionBufferUsed() const
     return true;
 }
 
-int CmdCardUpdateRecord::getSfi() const
+uint8_t CmdCardUpdateRecord::getSfi() const
 {
     return mSfi;
 }
 
-int CmdCardUpdateRecord::getRecordNumber() const
+uint8_t CmdCardUpdateRecord::getRecordNumber() const
 {
     return mRecordNumber;
 }
