+++ src/main/CmdCardUpdateRecord.h
@@ -56,7 +56,7 @@ public:
      */
     CmdCardUpdateRecord(const CalypsoCardClass calypsoCardClass,
                         const uint8_t sfi,
-                        const int recordNumber,
+                        const uint8_t recordNumber,
                         const std::vector<uint8_t>& newRecordData);
 
     /**
@@ -75,7 +75,7 @@ public:
      * @return The SFI of the accessed file
      * @since 2.0.1
      */
-    int getSfi() const;
+    uint8_t getSfi() const;
 
     /**
      * (package-private)<br>
@@ -83,7 +83,7 @@ public:
      * @return The number of the accessed record
      * @since 2.0.1
      */
-    int getRecordNumber() const;
+    uint8_t getRecordNumber() const;
 
     /**
      * (package-private)<br>
@@ -122,12 +122,12 @@ private:
     /**
      * Construction arguments
      */
-    const int mSfi;
+    const uint8_t mSfi;
 
     /**
      *
      */
-    const int mRecordNumber;
+    const uint8_t mRecordNumber;
 
     /**
      *
